+++ src/extension.ts
@@ -7,12 +7,14 @@ import { generateAuth } from "./commands/generateAuth";
 import { LoaderLens } from "./code-lenses/LoaderLens";
 import { ActionLens } from "./code-lenses/ActionLens";
 import { generateAuthSnippet } from "./commands/generateAuthSnippet";
+import { barrelize } from "./commands/barrelize";
 
 export function activate(context: vscode.ExtensionContext) {
   const flattenRoutesCommand = vscode.commands.registerCommand("remix-forge.flattenRoutes", flattenRoutes);
   const generateRouteFileCommand = vscode.commands.registerCommand("remix-forge.generateRemixRoute", generateRouteFile);
   const generateAuthCommand = vscode.commands.registerCommand("remix-forge.generateAuth", generateAuth);
   const openUrlCommand = vscode.commands.registerCommand("remix-forge.openUrl", openUrl);
+  const barrelizeCommand = vscode.commands.registerCommand("remix-forge.barrelize", barrelize);
   const generateAuthSnippetCommand = vscode.commands.registerCommand(
     "remix-forge.generateAuthSnippet",
     generateAuthSnippet
@@ -29,11 +31,14 @@ export function activate(context: vscode.ExtensionContext) {
   );
 
   // Register the commands
-  context.subscriptions.push(generateRouteFileCommand);
-  context.subscriptions.push(flattenRoutesCommand);
-  context.subscriptions.push(openUrlCommand);
-  context.subscriptions.push(generateAuthCommand);
-  context.subscriptions.push(generateAuthSnippetCommand);
+  context.subscriptions.push(
+    generateRouteFileCommand,
+    flattenRoutesCommand,
+    openUrlCommand,
+    generateAuthCommand,
+    generateAuthSnippetCommand,
+    barrelizeCommand
+  );
 }
 
 export function deactivate() {}